reverting search changes since they were buggin
