+++ app/search/page.tsx
@@ -1,3 +1,10 @@
+'use client';
+
+import { notFound } from 'next/navigation';
+import { Suspense } from 'react';
+import client from 'tina/__generated__/client';
+import { formatTableofContentsData } from 'utils/docs/getDocProps';
+import getTableOfContents from 'utils/docs/getTableOfContents';
 import SearchPageClient from './search-client';
 
 export async function generateMetadata() {
@@ -7,10 +14,55 @@ export async function generateMetadata() {
   };
 }
 
+async function getSearchPageData() {
+  const slug = 'index';
+  let tableOfContents, formatted;
+
+  try {
+    const results = await client.queries.doc({ relativePath: `${slug}.mdx` });
+    const documentData = results.data.doc;
+    tableOfContents = getTableOfContents(documentData.body.children);
+  } catch (e) {
+    console.error('Error Fetching Docs Navigation Data: ', e.message);
+    notFound();
+  }
+
+  try {
+    const query = `
+      query {
+        docsTableOfContents(relativePath: "docs-toc.json") {
+          _values
+        }
+      }
+    `;
+    const docTocData = await client.request(
+      {
+        query,
+        variables: { relativePath: 'docs-toc.json' },
+      },
+      {}
+    );
+    formatted = formatTableofContentsData(docTocData, null);
+  } catch (e) {
+    console.error('Error fetching Docs Table of Content Data: ', e.message);
+  }
+
+  return {
+    props: {
+      tableOfContents,
+      formatted,
+    },
+  };
+}
+
 export default async function SearchPage() {
+  const data = await getSearchPageData();
+
   return (
     <div>
-      <SearchPageClient />
+      <Suspense fallback={null}>
+        <SearchPageClient {...data} />
+      </Suspense>
     </div>
   );
 }
